+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 [Unreleased]: https://github.com/cashapp/redwood/compare/0.13.0...HEAD
 
 New:
-- Source-based schema parser is now the default. Can be disabled in your schema module with `redwood { useFir = false }`.
+- Source-based schema parser is now the default. The `useFir` Gradle property has been removed.
 - Introduce a `LoadingStrategy` interface to manage `LazyList` preloading.
 
 Changed: